Extract commands and events
